refactor solid and pipeline names, url inputs as list
