--- repositories/capturas/br_rj_riodejaneiro_sigmob/routes.yaml
@@ -1,10 +1,0 @@
-solids:
-  get_routes:
-    inputs:
-      url:
-        value: http://jeap.rio.rj.gov.br/MOB/get_routes.rule?sys=MOB
-resources:
-  basedosdados_config:
-    config:
-      table_id: routes
-      dataset_id: br_rj_riodejaneiro_sigmob
